moved target compile-junit into cpp2j.xml changed default location of base lib added generation of default c-tor changed some differences between structs and classes in meta.xml

+++ tests/t3_2/src/java/org/xbig/test/t3_2/BasicTests.java
@@ -16,7 +16,7 @@ public class BasicTests {
 	public void accessPublicStructAttribute() {
 		B a = new B();
 		a.setz(7.3);
-		int b = a.getz();
+		double b = a.getz();
 		Assert.assertEquals(7.3, b);
 		a.delete();
 	}
